Move javascript to their own files
